+++ src/condition/Any.js
@@ -6,5 +6,6 @@ module.exports = function () {
     var utils = require('../utils');
     
     return utils.customMessageWrapper(function anyCondition(value, path) {
+        return value;
     });
 };