+++ src/condition/Range.js
@@ -58,5 +58,7 @@ module.exports = function (options) {
             key = 'min' + (opts.minInclusive ? 'Inclusive' : '');
             throw new ConditionViolationException(value, path, messages[key], opts);
         }
+
+        return value;
     });
 };