+++ test/helper.js
@@ -7,9 +7,13 @@ var lewd = require('../src/lewd');
 function test(condition, args, accept, values) {
     values.forEach(function (value) {
         if (accept) {
+            var result;
+            
             buster.referee.refute.exception(function () {
-                condition.apply(lewd, args)(value);
+                result = condition.apply(lewd, args)(value);
             }, null, util.format('Value %s not expected to fail for arguments %s', value, args));
+            
+            buster.referee.assert.same(value, result);
         } else {
             buster.referee.assert.exception(function () {
                 condition.apply(lewd, args)(value);
@@ -22,11 +26,9 @@ module.exports = {
     acceptValues: function (condition, args, values) {
         test(condition, args, true, values);
     },
-    
     refuteValues: function (condition, args, values) {
         test(condition, args, false, values);
     },
-    
     refuteSchemaOptions: function (condition, args) {
         buster.referee.assert.exception(function () {
             condition.apply(lewd, args);